+++ src/main/java/com/example/stagemate/controller/ArchiveController.java
@@ -3,8 +3,8 @@ package com.example.stagemate.controller;
 import com.example.stagemate.domain.user.entity.UserJpaEntity;
 import com.example.stagemate.dto.request.ArchiveCreateRequest;
 import com.example.stagemate.dto.request.ArchiveUpdateRequest;
-import com.example.stagemate.dto.response.ArchiveDetailResponse;
-import com.example.stagemate.dto.response.ArchiveRankingResponse;
+import com.example.stagemate.dto.response.archive.ArchiveDetailResponse;
+import com.example.stagemate.dto.response.archive.ArchiveRankingResponse;
 import com.example.stagemate.global.dto.DataResponse;
 import com.example.stagemate.global.reslover.CurrentUser;
 import com.example.stagemate.service.archive.ArchiveService;
