+++ src/main/java/com/example/stagemate/controller/PerformanceController.java
@@ -3,8 +3,8 @@ package com.example.stagemate.controller;
 import com.example.stagemate.domain.performance.PerformanceGenre;
 import com.example.stagemate.domain.performance.PerformanceType;
 import com.example.stagemate.domain.user.entity.UserJpaEntity;
-import com.example.stagemate.dto.response.PerformanceDetailResponse;
-import com.example.stagemate.dto.response.RecommendedPerformanceResponse;
+import com.example.stagemate.dto.response.performance.PerformanceDetailResponse;
+import com.example.stagemate.dto.response.performance.RecommendedPerformanceResponse;
 import com.example.stagemate.global.dto.DataResponse;
 import com.example.stagemate.global.reslover.CurrentUser;
 import com.example.stagemate.service.performance.PerformanceService;
