+++ src/main/java/com/example/stagemate/service/archive/ArchiveService.java
@@ -6,8 +6,8 @@ import com.example.stagemate.domain.user.UserErrorCode;
 import com.example.stagemate.domain.user.entity.UserJpaEntity;
 import com.example.stagemate.dto.request.ArchiveCreateRequest;
 import com.example.stagemate.dto.request.ArchiveUpdateRequest;
-import com.example.stagemate.dto.response.ArchiveDetailResponse;
-import com.example.stagemate.dto.response.ArchiveRankingResponse;
+import com.example.stagemate.dto.response.archive.ArchiveDetailResponse;
+import com.example.stagemate.dto.response.archive.ArchiveRankingResponse;
 import com.example.stagemate.global.exception.AppException;
 import com.example.stagemate.global.exception.archive.ArchiveErrorCode;
 import com.example.stagemate.repository.ArchiveRepository;
